Rename pootle_pytest to pytest_pootle
Reasoning: this is a pytest plugin, not a Pootle plugin.
