+++ pytest_pootle/fixtures/core/utils/wordcount.py
