+++ pytest_pootle/fixtures/mock.py
