+++ pytest_pootle/fixtures/revision.py
