+++ pytest_pootle/fixtures/views.py
