+++ pytest_pootle/utils.py
