+++ setup.py
@@ -252,7 +252,7 @@ setup(
         'console_scripts': [
             'pootle = pootle.runner:main',
         ],
-        'pytest11': ['pootle = pootle_pytest.plugin'],
+        'pytest11': ['pootle = pytest_pootle.plugin'],
     },
     cmdclass={
         'build_checks_templates': BuildChecksTemplatesCommand,
