+++ tests/models/legalpage.py
@@ -9,11 +9,11 @@
 
 import pytest
 
+from pytest_pootle.factories import (AgreementFactory, LegalPageFactory,
+                                     UserFactory)
+
 from pootle.core.utils.timezone import aware_datetime
 from staticpages.models import LegalPage
-
-from pootle_pytest.factories import (
-    AgreementFactory, LegalPageFactory, UserFactory)
 
 
 @pytest.mark.django_db(transaction=True)
