+++ tests/models/user.py
@@ -14,13 +14,12 @@ import pytest
 
 from allauth.account.models import EmailAddress
 
+from pytest_pootle.fixtures.models.store import (TEST_EVIL_UPDATE_PO,
+                                                 _create_submission_and_suggestion,
+                                                 _create_comment_on_unit)
+
 import accounts
 from pootle_store.util import FUZZY, TRANSLATED
-
-from pootle_pytest.fixtures.models.store import (
-    TEST_EVIL_UPDATE_PO,
-    _create_submission_and_suggestion,
-    _create_comment_on_unit)
 
 
 def _make_evil_member_updates(store, evil_member):
